Add export methods for notice to NoticeExporter interface following Notice class refactor
- adapt import statements to said change
